+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Upcoming version 3.4.0 (????-??-??)
 
+### Added
+
+* Plugin has been tested with WordPress 6.6 [#111](https://github.com/chesio/bc-cache/issues/111).
+
 ### Changed
 
 * Button to flush the cache is now placed in site-related menu (left side of admin bar) and has a text label [#112](https://github.com/chesio/bc-cache/issues/112). This should improve UI/UX.