+++ CHANGELOG.md
@@ -7,7 +7,7 @@ This release has been tested with PHP 8.4.
 ### Added
 
 * Plugin has been tested with PHP 8.4 [#114](https://github.com/chesio/bc-security/issues/114).
-* Plugin has been tested with WordPress 6.6 [#111](https://github.com/chesio/bc-cache/issues/111).
+* Plugin has been tested with WordPress 6.6 [#111](https://github.com/chesio/bc-cache/issues/111) and 6.7 [#113](https://github.com/chesio/bc-cache/issues/113).
 
 ### Changed