+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # BC Cache Changelog
 
-## Upcoming version 3.4.0 (????-??-??)
+## Version 3.4.0 (2024-10-28)
 
 This release has been tested with PHP 8.4.